pyproject.toml
[tool.poetry]
name = "ukw-detection-system"
version = "0.1.0"
description = ""
authors = ["martineserios <mrios90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
scikit-learn = "^1.4.1.post1"
python-dotenv = "^1.0.1"
ipytest = "^0.14.0"
tensorflow = "2.14"
sagemaker = {extras = ["local"], version = "^2.214.3"}
ruff = "^0.3.5"
nbformat = "^5.10.4"
nbdev = "^2.3.13"
nbclient = "^0.10.0"
ipykernel = "^6.29.4"
flask = "^3.0.3"
matplotlib = "^3.8.4"
numpy = "^1.26.4"
pillow = "^10.3.0"
psutil = "^5.9.8"
pyyaml = "^6.0.1"
requests = "^2.31.0"
scipy = "^1.13.0"
thop = "^0.1.1.post2209072238"
torch = "^2.2.2"
torchvision = "^0.17.2"
tqdm = "^4.66.2"
ultralytics = "^8.1.45"
gitpython = "^3.1.43"
boto3 = "^1.34.107"
opencv-python = "4.8.0.74"
twilio = "^9.0.5"
loguru = "^0.7.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"